terminal-transmitter/pyproject.toml
[tool.poetry]
name = "terminal-transmitter"
version = "0.1.0"
description = "Микросервис запуска модели"
authors = ["wisata"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
fastapi = "~0.104.1"
faststream = {extras = ["rabbit"], version = "^0.3.14"}
uvicorn = "^0.29.0"
redis = "^5.0.4"
python-socketio = "^5.11.2"
boto3 = "^1.34.124"
pillow = "^10.3.0"
ultralytics = "^8.2.32"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
